Webfunc-python310-dify/src/requirements.txt
cloudbase-agent-core==0.1.11
cloudbase-agent-server==0.1.16
cloudbase-agent-dify==0.1.0b1
httpx>=0.24.0
python-dotenv>=1.2.1
